In src/list.scm rac, rdc, snoc and repeated added.
In src/axiom.scm
imp-formulas-to-uninst-gfp-formulas-etc changed to use OrNc for nc-coidpcs.

In src/pconst.scm
corec-const-and-bound-to-bcorec-term changed to use partial variables
instead of total variables in order to form unfolded terms.
const-substitute changed to support CoRec.

In src/psym.scm
arity-to-new-harrop-pvar and nc-idpredconst-name? added.  add-ids
changed to use harrop-pvars in a proper way.  add-mr-ids changed to
deal with computational content by means of nc as well as cr
inductively defined conjunction and existential quantifier.

In src/proof.scm
and-formula-and-concl-to-and-elim-proof added to provide elimination
proof for primitive conjunction.

In src/formula.scm
exi-form?, exi-mr-exi-form?, exi-mr-exi-form-to-vars,
exi-mr-exi-form-to-final-kernel,
and-andi-ex-exi-formula-to-vars-and-conjuncts, ori-mr-ori-form?,
andi-form?, andi-mr-andi-form?, andi-mr-andi-form-to-conjuncts,
and-andi-mr-andi-form-to-conjuncts and eqd-form? added for
convenience.

In src/ets.scm
axiom-to-extracted-term changed to take tpsubst into account.
coidpredconst-to-closure-mr-proof, imp-formulas-to-mr-gfp-proof,
term-to-eqd-proof,
alg-or-arrow-types-to-unfolded-corec-eqd-global-assumption-name, and
all-allnc-form-and-prems-and-opt-goal-fla-to-unifier added.  Also some
helper procedures imp-formulas-to-mr-gfp-proof-aux,
str-gfp-proof-helper, coidpredconst-to-closure-mr-proof-or-intro, and
coidpredconst-to-closure-mr-proof-or-elim added.

In examples/test/ets.scm
Examples added to test coidpredconst-to-closure-mr-proof and
imp-formulas-to-mr-gfp-proof.