EqP EqPNc CoEqP CoEqPNc EqPMR CoEqPMR added as predconsts.  In formula-to-et-type-for-mr-clauses EqP and CoEqP added.  add-eqp with auxiliary functions alg-name-to-eqp-idpredconst-name alg-to-eqp-idpredconst term-and-alist-to-compatibility-formula term-to-eqp-compatibility-formula added.  Similarly: add-eqpnc added.
In add-mr-ids application to totality-idpcs forbidden, since
TotalAlgNc suffices.  In real-and-formula-to-mr-formula-aux usage of
TotalAlgMR replaced by TotalAlgNc.  TotalNc CoTotal CoTotalNc TotalMR
EqP EqPNc CoEqP CoEqPNc EqPMR CoEqPMR added as predconsts.  In
formula-to-et-type-for-mr-clauses CoTotal EqP CoEqP added.
add-totalnc with auxiliary function
alg-name-to-totalnc-idpredconst-name alg-to-totalnc-idpredconst
term-and-alist-to-totalnc-formula added.  add-rtotalnc with auxiliary
functions alg-name-to-rtotalnc-idpredconst-name
alg-to-rtotalnc-idpredconst term-and-alist-to-rtotalnc-formula added.

add-co now takes an optional argument opt-prim-prod-eq-info.  If 'prim
is present clauses are formed with mk-ex, mk-and (generating extracted
terms with the primitive product make-star).  Otherwise the clauses
are formed with the inductively defined existential quantifiers and
conjunctions (generating extracted terms with the defined product
yprod).  The other items are lists of eq-names, usually as many as
there are (simultaneously defined) idpc-names.  Each list should have
as many eq-names as there are arguments, with fitting types.  Too
short lists are filled by EqD, and missing lists are taken to be lists
of EqD.  For clarity clause-to-exand-formula taken out of add-co .
c-flag removed.  It is determined by nc-idpc?
eq-name-to-predicate-generator added.