formula-to-et-type formula-of-nulltype? and formula-of-nulltype-under-extension? extended to CoTotal EqP CoEqP andlr-formula-and-concl-to-andlr-elim-mr-proof corrected (work of Kenji Miyamoto).
Because of the change of real-and-formula-to-mr-formula for totality
formulas we redefine invarex-aconst invarall-aconst and remove
totalmrtoeq-aconst and the theorems InvarEx InvarAll TotalMRToEq
TotalMRIntro TotalMRToTotal.
real-and-formula-to-invar-proof added.
Given a term t and a formula A, real-and-formula-to-invar-proof
returns a proof of t mr A -> A.  There can be parameters in t,A.

There is another way to introduce and eliminate the abbreviation allnc
alpha with a total variable alpha, which uses TotalNc alpha^ -> rather
than Total alpha^ --> .  Theorems TotalToTotalNc TotalNcToTotal
AllncTotalIntroTotalNc AllncTotalElimTotalNc TotalNcIntro
CoTotalToCoTotalNc CoTotalNcToCoTotal added.

In theorems ExRTotalElim AllTotalIntroSound AllTotalElimSound usage
of TotalMR eliminated.